Diffusion map implementation now works
The UnitRangeTransform is not available in the Stats packages yet,
so I rewrote my own quick and dirty one. That should probably be
replaced with the more robust transform later.

The distance calculation was behaving weirdly for me, so I replaced
it with a calculation from Distances.jl. This also need a requirement
for Distances added to REQUIRE.

I renamed the argument t to α to keep with what I've seen in the
Coifman and Lafon articles. That's the parameter which defines whether
you are using the weighted graph Laplacian or approximate Laplace-Beltrami
operator, not really the time step of the diffusion process.

The tests were modified to account for the slight API change, and the
main module was modified to add the dependency on Distances.jl